requirements.txt
streamlit>=1.32.0
mediapipe>=0.10.0
opencv-python-headless>=4.8.0
openai>=1.0.0
pillow>=10.0.0
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.24.0
